Reorganize frontend code structure, add navigation to hufu and support i18n
