+++ webapp/babel.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+  compact: false,
   presets: [
     '@vue/cli-plugin-babel/preset'
   ]
