+++ webapp/src/assets/i18n/languages/en_US.js
@@ -1,0 +1,14 @@
+module.exports =  {
+    language: {
+        locale: "Language",
+        chinese: "Chinese",
+        english: "English"
+    },
+    menu: {
+        overview: "Overview",
+        editor: "Editor",
+        operation: "Operation",
+        datasource: "Datasource",
+        setting: "Setting"
+    },
+}
