+++ webapp/src/assets/i18n/languages/zh_CN.js
@@ -1,0 +1,14 @@
+module.exports =  {
+    language: {
+        locale: "语言",
+        chinese: "中文",
+        english: "英文"
+    },
+    menu: {
+        overview: "概览",
+        editor: "SQL编辑器",
+        operation: "操作",
+        datasource: "数据源",
+        setting: "设置"
+    },
+}
