+++ webapp/src/pages/datasource/datasource.vue
@@ -1,0 +1,13 @@
+<template>
+  <router-view></router-view>
+</template>
+
+<script>
+
+export default {
+  name: 'DatasourcePage',
+  components: {
+    // eslint-disable-next-line
+  },
+}
+</script>
