+++ webapp/src/pages/editor/editor-router.js
@@ -1,0 +1,14 @@
+import {createRouter, createWebHistory} from 'vue-router'
+
+import EditorMain from './editor-main';
+
+export default createRouter({
+    history: createWebHistory(),
+    routes: [
+        {
+            path: '/',
+            name: 'editor',
+            component: EditorMain
+        }
+    ]
+});
