+++ webapp/src/pages/operation/operation.vue
@@ -1,0 +1,13 @@
+<template>
+  <router-view></router-view>
+</template>
+
+<script>
+
+export default {
+  name: 'OperationPage',
+  components: {
+    // eslint-disable-next-line
+  },
+}
+</script>
