+++ webapp/src/pages/overview/overview.vue
@@ -1,0 +1,13 @@
+<template>
+  <router-view></router-view>
+</template>
+
+<script>
+
+export default {
+  name: 'OverviewPage',
+  components: {
+    // eslint-disable-next-line
+  },
+}
+</script>
