+++ webapp/src/pages/setting/setting.vue
@@ -1,0 +1,13 @@
+<template>
+  <router-view></router-view>
+</template>
+
+<script>
+
+export default {
+  name: 'SettingPage',
+  components: {
+    // eslint-disable-next-line
+  },
+}
+</script>
