--- webapp/src/router/router.js
@@ -1,11 +1,0 @@
-import { createRouter, createWebHashHistory } from "vue-router";
-
-const routes = [
-];
-
-const router = createRouter({
-  history: createWebHashHistory(),
-  routes,
-});
-
-export default router;
