+++ webapp/vue.config.js
@@ -1,3 +1,6 @@
+var webpack=require('webpack')
+
+const { VueLoaderPlugin } = require('vue-loader')
 const { defineConfig } = require("@vue/cli-service");
 module.exports = defineConfig({
   transpileDependencies: true,
@@ -7,5 +10,17 @@ module.exports = defineConfig({
         config.output.filename("background.js");
       },
     },
+  },
+  configureWebpack: config => {
+    if (process.env.NODE_ENV === 'production') {
+      config.mode = 'production';
+      config.performance = {
+        maxEntrypointSize: 10000000,
+        maxAssetSize: 30000000
+      }
+    }
+  },
+  devServer: {
+    port: 8080
   },
 });
